Рефакторинг: объединяет компонент controls с info-section
